Move namespaces JSON test to test_stashcache.py so I can use the test global data
keep a simple endpoint test in test_api.py
